Parte 2

+++ Backend/Infrastructure/Command/DishCommand.cs
@@ -1,41 +1,44 @@
-﻿using Application.Interfaces.IDish;
+using Application.Interfaces.IDish;
 using Domain.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;   
 using Infrastructure.Data;
 using Microsoft.EntityFrameworkCore;
-
+using System;
+using System.Threading.Tasks;
 
 namespace Infrastructure.Command
 {
     public class DishCommand : IDishCommand
     {
         private readonly AppDbContext _context;
+        
         public DishCommand(AppDbContext context)
         {
             _context = context;
         }
 
-        public async Task InsertDish(Dish dish)
+        public async Task<Dish> CreateDish(Dish dish)
         {
             _context.Dishes.Add(dish);
             await _context.SaveChangesAsync();
+            return dish;
         }
 
-        public async Task RemoveDish(Dish dish)
-        {
-            _context.Dishes.Remove(dish);
-            await _context.SaveChangesAsync();
-        }
-
-        public async Task UpdateDish(Dish dish)
+        public async Task<Dish> UpdateDish(Dish dish)
         {
             _context.Dishes.Update(dish);
             await _context.SaveChangesAsync();
+            return dish;
+        }
+
+        public async Task<bool> DeleteDish(Guid id)
+        {
+            var dish = await _context.Dishes.FindAsync(id);
+            if (dish == null)
+                return false;
+            
+            _context.Dishes.Remove(dish);
+            await _context.SaveChangesAsync();
+            return true;
         }
     }
 }
-
